Allow to compile GlideHQ as part of mupen64plus-video-glide64mk2

+++ src/GlideHQ/Ext_TxFilter.h
@@ -48,8 +48,9 @@
 typedef __int64 int64;
 typedef unsigned __int64 uint64;
 #else
-typedef long long int64;
-typedef unsigned long long uint64;
+#include <stdint.h>
+typedef int64_t int64;
+typedef uint64_t uint64;
 typedef unsigned char boolean;
 #endif
